requirements.txt
django
psycopg2
django-debug-toolbar
django-geojson [field]
django-nap